3. Optional в банковских переводах. [#214711]4

+++ chapter_003/src/main/java/ru/job4j/bank/Account.java
@@ -28,7 +28,7 @@ public class Account {
     public String getRequisites() {
         return this.requisites;
     }
-    public Account(int value, String requisites) {
+    public Account(double value, String requisites) {
         this.value = value;
         this.requisites = requisites;
     }
